test: split CPU and GPU test paths
