+++ scripts/run_cpu_tests.sh
@@ -22,20 +22,65 @@ run_step \
     cargo test -p cutile-compiler --doc
 
 run_step \
+    "cutile-compiler Tile IR to cubin lowering tests" \
+    cargo test -p cutile-compiler --test compiler2_e2e
+
+run_step \
     "cutile library tests" \
     cargo test -p cutile --lib
 
 run_step \
-    "cutile compile-only FMHA regression" \
-    cargo test -p cutile --test flash_attention_compile
+    "cutile doc tests" \
+    cargo test -p cutile --doc
+
+for test_target in \
+    basics_and_inlining \
+    binary_math_ops \
+    bitwise_and_bitcast_ops \
+    compile_error_quality \
+    compile_only \
+    element_type_zero \
+    element_type_zero_jit \
+    error_quality \
+    flash_attention_compile \
+    global_memory \
+    integer_ops \
+    kernel_compiler \
+    load_tile_like_examples \
+    macro_submodule \
+    memory_and_atomic_ops \
+    num_tiles_jit \
+    optimization_hints \
+    partition_index_schedules \
+    reduce_scan_ops \
+    registry_phase_a \
+    span_source_location \
+    trait_dispatch_probe \
+    two_cga_trait_impl \
+    type_inference_sanity \
+    unary_math_ops \
+    use_classifier
+do
+    run_step \
+        "cutile CPU integration test ${test_target}" \
+        cargo test -p cutile --test "$test_target"
+done
 
 run_step \
-    "cutile load_tile_like compile-only regressions" \
-    cargo test -p cutile --test load_tile_like_examples
+    "cutile compile-only control-flow regressions" \
+    cargo test -p cutile --test control_flow_ops compile_
 
 run_step \
-    "cutile type inference sanity regressions" \
-    cargo test -p cutile --test type_inference_sanity
+    "cutile compile-only tensor and matrix regressions" \
+    cargo test -p cutile --test tensor_and_matrix_ops compile_
+
+run_step \
+    "cutile compile-only type conversion regressions" \
+    cargo test -p cutile --test type_conversion_ops compile_
+
+run_step \
+    "cutile compile-only specialization regressions" \
+    cargo test -p cutile --test specialization_bits -- --skip raw_pointer_launch
 
 print_summary_and_exit \
     "All CPU tests passed!" \
